+++ src/class.js
@@ -66,22 +66,7 @@ var Class = Object.create(null, {
   createStatic: {
     value: function (properties) {
       var obj = {};
-      TypeBuilder.forEach(properties, function (name, value) {
-        if (TypeBuilder.isConstructor(name)) {
-          TypeBuilder.addConstructor(obj, name, value);
-        } else if (TypeBuilder.isMethod(value)) {
-          TypeBuilder.addMethod(obj, name, value);
-        } else if (TypeBuilder.isProperty(value)) {
-          TypeBuilder.addProperty(obj, name, value["get"], 
-            value["set"]);
-        } else if (TypeBuilder.isConstant(name)) { 
-          TypeBuilder.addConstant(obj, name, value); 
-        } else if (TypeBuilder.isStatic(name)) {
-          TypeBuilder.addStatic(obj, value); 
-        } else {
-          TypeBuilder.addField(obj, name, value);
-        }      
-      });
+      TypeBuilder.addStatic(obj, properties);
       Class.__initialize.call(obj, obj);
       Object.seal(obj);
       return obj; 
@@ -309,6 +294,7 @@ var Class = Object.create(null, {
           TypeBuilder.addField(constructor.prototype, name, value);
         }
       });
+      Class.__initialize.call(constructor.prototype, constructor.prototype);
       Object.seal(constructor);
       Object.seal(constructor.prototype);
       if (interfaces) {